将第八题添加到Form1代码中，简化Form1代码，测试方法和测试版本会随着测试问题变化而改变，增加了一些结果文件，修改SaleSystem测试文件中变量类型错位问题

+++ Test/Test/Tests/SaleSystem/SaleSystemDicisionConditionTest.cs
@@ -16,7 +16,7 @@ namespace Test.Tests
 		public double CalculateAmout()
 		{
 			int attribute = 0;  // 佣金系数
-		    int result = 0;  // 佣金
+		    double result = 0;  // 佣金
 		    if(amount > 200 && freeDay <= 10){
 		        if(cash >= 60) {
 		            attribute = 7;
